+++ work_spacer/lib/models/desk.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/foundation.dart';
+
+@immutable
+class Desk {
+  final int id;
+  final int floor;
+  final bool secondMonitor;
+
+  const Desk({
+    required this.id,
+    required this.floor,
+    required this.secondMonitor,
+  });
+}