+++ work_spacer/lib/models/reservation.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/foundation.dart';
+
+@immutable
+abstract class Reservation {
+  final DateTime startDate;
+  final int duration;
+  final int idEmployee;
+
+  const Reservation({
+    required this.startDate,
+    required this.duration,
+    required this.idEmployee,
+  });
+}