+++ work_spacer/lib/screens/admin/roles/role_management_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class RoleManagementScreen extends StatelessWidget {
+  static const routeName = '/roles';
+
+  const RoleManagementScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Manage roles'),
+      ),
+      body: Container(),
+    );
+  }
+}