+++ work_spacer/lib/screens/employee/desks/desk_search_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class DeskSearchScreen extends StatelessWidget {
+  static const routeName = '/desks';
+
+  const DeskSearchScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Desks'),
+      ),
+      body: Container(),
+    );
+  }
+}