+++ work_spacer/lib/screens/employee/notifications/notifications_screen.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class NotificationsScreen extends StatelessWidget {
+  static const routeName = '/notifications';
+
+  const NotificationsScreen({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Notifications'),
+      ),
+      body: Container(),
+    );
+  }
+}