+++ work_spacer/lib/screens/employee/rooms/room_search_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class RoomSearchScreen extends StatelessWidget {
+  static const routeName = '/rooms';
+
+  const RoomSearchScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Rooms'),
+      ),
+      body: Container(),
+    );
+  }
+}