+++ CHANGELOG.md
@@ -10,3 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 * Backend template (strapi)
 * App template (flutter)
+
+## [0.1.0] - 2022-05-23
+
+### Added
+* Created all the screens without content, with navigation.
+* Prepared the models.
+* Created a KeyboardHideWrapper. Use it to wrap a screen that has a textfield inside. Now if you click outside, the textfield will no longer stay focused.
+
+### Changed
+* SDK version changed from 2.16.1 to 2.17.0 in order to use new syntax for super constructors.